// src/entities/openapi/model/documentScanStatus.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const DocumentScanStatus = {
    PendingScan: 'PENDING_SCAN',
    Available: 'AVAILABLE',
    Rejected: 'REJECTED',
    ScanFailed: 'SCAN_FAILED'
} as const;
export type DocumentScanStatus = typeof DocumentScanStatus[keyof typeof DocumentScanStatus];